requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
breadability==0.1.20
cachetools==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
colorama==0.4.6
decorator==5.1.1
Deprecated==1.2.14
docopt==0.6.2
extra-streamlit-components==0.1.56
fasttext-wheel==0.9.2
filelock==3.12.2
Flask==2.3.2
Flask-Limiter==3.3.1
frozenlist==1.3.3
fsspec==2023.6.0
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.59.0
hazm @ git+https://github.com/roshan-research/hazm.git@446946bd0dc84453add7722cf4cbabd38fa92d57
httplib2==0.22.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
libwapiti==0.2.1
limits==3.5.0
lit==16.0.6
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.7
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycountry==22.3.5
pydeck==0.8.1b0
pyfiglet==0.8.post1
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
python-resize-image==1.1.20
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rake-nltk==1.0.6
regex==2023.6.3
requests==2.31.0
rich==13.4.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.11.2
six==1.11.0
smart-open==6.3.0
smmap==5.0.0
streamlit==1.25.0
streamlit-authenticator==0.2.2
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
summa==1.2.0
sumy==0.11.0
sympy==1.12
tenacity==8.2.2
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.6
wrapt==1.15.0
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
